openfaas-functions/image-recognition/requirements.txt
tensorflow==2.10.0
pillow==9.5.0
numpy==1.24.3 